moved filedata tests in test package
